microservices/bUploadPictureFromCamera/requirements.txt
python-dotenv==0.10.3
imgcompare==2.0.1
google-cloud-pubsub==1.0.2
google-cloud-storage==1.23.0
google-cloud-logging==1.14.0